# laravel_app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "KOPOKOPO Transactions",
        "description": "Kopokopo Transactions Management",
        "contact": {
            "email": "martinwainaina001@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/accounts/{account_id}": {
            "get": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Get a transaction by account_id",
                "operationId": "44ae4ce99c0ab35e2ecfc19af91907ea",
                "parameters": [
                    {
                        "name": "account_id",
                        "in": "path",
                        "description": "ID of the account",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "0afd02d3-6c59-46e7-b7bc-893c5e0b7ac2"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Account_id retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Account_id retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "account_id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "0afd02d3-6c59-46e7-b7bc-893c5e0b7ac2"
                                                },
                                                "balance": {
                                                    "type": "number",
                                                    "example": 17
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Account_id not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Account not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}